Extract ingress instructions into separate component
