+++ src/components/Cluster/ClusterDetail/Ingress/Ingress.tsx
@@ -1,50 +1,21 @@
 import styled from '@emotion/styled';
+import Instructions from 'Cluster/ClusterDetail/Ingress/Instructions';
 import PropTypes from 'prop-types';
 import React from 'react';
 import { Providers } from 'shared/constants';
 import { PropertiesOf } from 'shared/types';
 
-import {
-  Description,
-  Emphasis,
-  Info,
-  InfoRow,
-  StyledURIBlock,
-  Text,
-  URIWrapper,
-} from './Components';
-import Option from './Option';
-import Ports from './Ports';
+import { Emphasis, Text } from './Components';
 
 const IngressWrapper = styled.div`
   max-width: 1200px;
 
   & > .col-12 > .row + .row {
-    margin-top: 16px;
+    margin-top: ${({ theme }) => theme.spacingPx * 4}px;
   }
 `;
 
-const Label = styled.span`
-  flex: 0 0 220px;
-  max-width: 100%;
-`;
-
-const Steps = styled.ol`
-  padding-left: 0;
-  list-style-position: inside;
-
-  ${URIWrapper} {
-    margin-top: 8px;
-  }
-`;
-
-enum IngressPathPrefixes {
-  Pattern = 'YOUR_PREFIX',
-  LoadBalancer = 'ingress',
-  Example = 'example',
-}
-
-interface IIngressProps {
+interface IIngressProps extends React.ComponentPropsWithoutRef<'div'> {
   provider?: PropertiesOf<typeof Providers>;
   k8sEndpoint?: string;
   kvmTCPHTTPPort?: number;
@@ -60,12 +31,6 @@ const Ingress: React.FC<IIngressProps> = ({
   hasOptionalIngress,
   ...rest
 }) => {
-  // Safe because of default props
-  const basePath: string = getBasePathFromK8sEndpoint(k8sEndpoint as string);
-  const examplePath: string = `${IngressPathPrefixes.Example}${basePath}`;
-  const balancerPath: string = `${IngressPathPrefixes.LoadBalancer}${basePath}`;
-  const patternPath: string = `${IngressPathPrefixes.Pattern}${basePath}`;
-
   return (
     <IngressWrapper className='row' {...rest}>
       <div className='col-12'>
@@ -82,113 +47,13 @@ const Ingress: React.FC<IIngressProps> = ({
             </Text>
           </div>
         </div>
-        <div className='row'>
-          <div className='col-12'>
-            <Info>
-              <InfoRow>
-                <Label>Base domain:</Label>
-                <URIWrapper>
-                  <StyledURIBlock>{basePath}</StyledURIBlock>
-                </URIWrapper>
-              </InfoRow>
-              <InfoRow>
-                <Label>Load balancer DNS name:</Label>
-                <URIWrapper>
-                  <StyledURIBlock>{balancerPath}</StyledURIBlock>
-                </URIWrapper>
-              </InfoRow>
-              <InfoRow>
-                <Label>Hostname pattern:</Label>
-                <URIWrapper>
-                  <StyledURIBlock copyContent={patternPath}>
-                    <Emphasis>{IngressPathPrefixes.Pattern}</Emphasis>
-                    {basePath}
-                  </StyledURIBlock>
-                  <Description>
-                    Replace <code>{IngressPathPrefixes.Pattern}</code> with a
-                    unique domain name segment to address your service.
-                  </Description>
-                </URIWrapper>
-              </InfoRow>
-
-              {provider === Providers.KVM && (
-                <InfoRow>
-                  <Label>Load balancer TCP ports:</Label>
-                  <Ports HTTP={kvmTCPHTTPPort} HTTPS={kvmTCPHTTPSPort} />
-                </InfoRow>
-              )}
-            </Info>
-          </div>
-        </div>
-        <div className='row'>
-          <div className='col-12'>
-            <Text>
-              To expose services in this cluster to the internet, configure your
-              ingress resources by choosing one of the following options (A) or
-              (B):
-            </Text>
-          </div>
-        </div>
-        <div className='row'>
-          <div className='col-12'>
-            <Option
-              id='a'
-              footer={
-                <>
-                  <Text>
-                    The benefit here is that if the service moves to a different
-                    cluster, the CNAME can be modified to point to the new
-                    cluster&apos;s load balancer DNS.
-                  </Text>
-                  <Text />
-                  Caution: DNS change propagation may take up to a day,
-                  depending on the TTL configuration of the DNS entry.
-                </>
-              }
-            >
-              <Steps>
-                <li>
-                  Set the <code>hostname</code> in your workload&apos;s ingress
-                  resource to a custom DNS name like e. g.{' '}
-                  <code>myservice.example.com</code>.
-                </li>
-                <li>
-                  Create a CNAME <code>myservice.example.com</code> pointing to
-                  the load balancer DNS name
-                  <URIWrapper>
-                    <StyledURIBlock>{balancerPath}</StyledURIBlock>
-                  </URIWrapper>
-                </li>
-              </Steps>
-            </Option>
-            <Option
-              id='b'
-              footer={
-                <Text>
-                  This option does not require dealing with DNS configuration.
-                  The downside is that when the service moves to a different
-                  cluster, clients consuming the service can no longer use the
-                  known URL and will need the new one.
-                </Text>
-              }
-            >
-              <Steps>
-                <li>
-                  Define a DNS prefix unique within the cluster, e. g.{' '}
-                  <code>example</code>.
-                </li>
-                <li>
-                  Set the <code>hostname</code> in your workload&apos;s Ingress
-                  resource to the according name
-                  <URIWrapper>
-                    <StyledURIBlock>{examplePath}</StyledURIBlock>
-                  </URIWrapper>
-                </li>
-              </Steps>
-            </Option>
-          </div>
-        </div>
       </div>
+      <Instructions
+        provider={provider}
+        k8sEndpoint={k8sEndpoint}
+        kvmTCPHTTPPort={kvmTCPHTTPPort}
+        kvmTCPHTTPSPort={kvmTCPHTTPSPort}
+      />
     </IngressWrapper>
   );
 };
@@ -210,14 +75,3 @@ Ingress.defaultProps = {
 };
 
 export default Ingress;
-
-function getBasePathFromK8sEndpoint(endpoint: string): string {
-  const replacer = 'https://api';
-  let result = '';
-
-  if (endpoint.startsWith(replacer)) {
-    result = endpoint.replace(replacer, '');
-  }
-
-  return result;
-}
